pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fnuxxmlparser"
version = "0.1.0"
description = "FNUX XML Parser for medical data processing and summarization"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Ole" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Markup :: XML",
]
dependencies = [
    "openai>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.urls]
Documentation = "https://github.com/username/fnuxxmlparser#readme"
Source = "https://github.com/username/fnuxxmlparser"

[project.scripts]
fnuxxmlparser = "fnuxxmlparser.parser:main"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=fnuxxmlparser"
testpaths = [
    "tests",
]
